setup test for username
