+++ types/index.go
@@ -1,0 +1,6 @@
+package types
+
+type User struct {
+	Email    string
+	Username string
+}
